CI-790 9th Feb commit part way through working out the page flows for search

+++ src/Web/Sfa.Das.Sas.Core/Domain/Model/ProviderSearchResultSummary.cs
@@ -2,7 +2,7 @@
 {
     public class ProviderSearchResultSummary
     {
-        public int UkPrn { get; set; }
+        public long UkPrn { get; set; }
         public string ProviderName { get; set; }
     }
 }
